requirements.txt
fastapi==0.110.0
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.3
pydantic==2.6.4
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.1
python-dotenv==1.0.1